requirements/base.txt
Django==5.2.1
django-admin-sortable2==2.2.8
django-auditlog==3.1.2
django-cors-headers==4.7.0
django-filter==25.1
django-import-export==4.3.7
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
google-auth-oauthlib==1.2.2
google-api-python-client==2.169.0
mysql-connector-python==9.3.0
pillow==11.2.1
python-decouple==3.8
whitenoise==6.9.0